Submission of kissmig
post_processing: From a given origin or initial distribution the spread/shrinking pattern of a given species is simulated on top of pre-processed SDM suitability maps. Changing environments can be considers by a stack of different suitability maps. KISSMig provides as output maps of the final species distribution (option ’DIS’), the iteration step of first occurrence (’FOC’), the iteration step of last occurrence (’LOC’), or the number of iteration steps with occurrence (’NOC’). Iteration steps can be interpreted as time.